phase3_introducing_generative_ai/services/shared/requirements.txt
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-requests==0.46b0
Flask==3.0.3
requests==2.32.3
prometheus-flask-exporter==0.23.0
setuptools>=68,<70
wheel
